backend/requirements.txt
torch
torchvision
open_clip_torch
opencv-python
pillow

chromadb
google-genai

fastapi
uvicorn
sqlalchemy
pyttsx3
python-jose[cryptography]
python-multipart
pydantic[email]
passlib[argon2]
python-dotenv
email-validator
